Ref : Ajout test existence fichier avant suppression.
git-svn-id: svn://svn.code.sf.net/p/mediboard/code/trunk@18612 a21d6be2-e512-0410-a060-91f5a2946631

+++ classes/CMbLock.class.php
@@ -55,6 +55,10 @@ class CMbLock {
    * @return bool
    */
   function release() {
-    return unlink($this->lock_file);
+    if (file_exists($this->lock_file)) {
+      return unlink($this->lock_file);
+    }
+
+    return true;
   }
 }
